+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
  Usagi Changelog
 =================
 
-Changes since version 0.1.1
-===========================
+Version 0.2.0
+=============
 
 Enhancements
 ------------